refactor(dto): util 패키지의 위치를 가장 상위로 이동
