+++ src/main/java/util/StringUtil.java
@@ -1,0 +1,8 @@
+package util;
+
+public class StringUtil {
+
+    public static String[] split(String input, String regex) {
+        return input.split(regex, -1);
+    }
+}
